refactor: refactor folder structure for build fix
