+++ src/app/(main)/ticker/_components/content.tsx
@@ -1,14 +1,10 @@
-"use client"
-
 import TickerList from "./ticker-list";
 
 const Content = () => {
   const handleClickTicker = (name: string) => {
     console.log("name:", name);
-  }
-  return (
-    <TickerList onClick={handleClickTicker}/>
-  );
+  };
+  return <TickerList onClick={handleClickTicker} />;
 };
 
-export default Content
+export default Content;
