+++ src/app/(main)/ticker/_components/intro.tsx
