+++ src/app/(main)/ticker/_components/ticker-drawer.tsx
