+++ src/app/(main)/ticker/page.tsx
