--- src/components/common/Logo.tsx
@@ -1,7 +1,0 @@
-import React from "react";
-
-const Logo = () => {
-  return <p className="px-1 text-lg font-bold">Logo입니다</p>;
-};
-
-export default Logo;
