+++ src/components/common/lottie-player/index.ts
