+++ src/components/common/lottie-player/lottie-player.tsx
